+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fix
 
+- Export: Add block_index if export blocks
 - PhysicalObject : Remove to_stl encoding element
 - Schema : Take `None` default value into account
 - WorkflowRun : Compute memory usage, subtract before_memory